Refactored project structure to improve modularity and prevent circular imports
