+++ quickannotator/constants.py
@@ -14,3 +14,5 @@ MASK_CLASS_ID = 1
 TILE_PRED_EXPIRE = 1 # minutes
 
 MAX_ACTORS_PROCESSING = 1
+
+FLASK_DATETIME_FORMAT = 'iso'
