+++ quickannotator/db/crud/misc.py
@@ -1,0 +1,2 @@
+def compute_custom_metrics() -> dict:
+    return {"iou": 0.5}
