+++ quickannotator/db/crud/project.py
@@ -1,9 +1,9 @@
+import quickannotator.db.models as db_models
 from quickannotator.db import db_session
-import quickannotator.db.models as models
-from datetime import datetime
+
 
 def add_project(name, description, is_dataset_large):
-        project = models.Project(name=name,
+        project = db_models.Project(name=name,
                           description=description,
                           is_dataset_large=is_dataset_large)
         db_session.add(project)
