+++ quickannotator/tests/conftest.py
@@ -5,10 +5,10 @@ from quickannotator.db import init_db, drop_db, db_session, get_session
 from quickannotator.api import init_api
 from quickannotator.config import get_database_uri, get_api_version
 from quickannotator.db import models
-from quickannotator.api.v1.project.utils import add_project
-from quickannotator.api.v1.image.utils import add_image_by_path
-from quickannotator.api.v1.annotation_class.helper import insert_annotation_class
-from quickannotator.api.v1.utils.shared_crud import AnnotationStore
+from quickannotator.db.crud.project import add_project
+from quickannotator.db.crud.image import add_image_by_path
+from quickannotator.db.crud.annotation_class import insert_annotation_class
+from quickannotator.db.crud.annotation import AnnotationStore
 from quickannotator.db.crud.tile import TileStore, TileStoreFactory
 from quickannotator.constants import TileStatus
 from shapely.geometry import Polygon
